third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger/network-uisourcecode-provider.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/debugger-test.js"></script>
<script>
function addIframe()
{
    var iframe = document.createElement("iframe");
    iframe.setAttribute("src", "resources/syntax-error.html");
    document.body.appendChild(iframe);
}

function addScript()
{
    var script = document.createElement("script");
    script.setAttribute("src", "resources/script1.js");
    document.head.appendChild(script);
}

function addStyleSheet()
{
    var style = document.createElement("link");
    style.setAttribute("rel", "stylesheet");
    style.setAttribute("type", "text/css");
    style.setAttribute("href", "resources/style1.css");
    document.head.appendChild(style);
    window._style = style;
}

function removeStyleSheet()
{
    window._style.parentNode.removeChild(window._style);
    window._style = null;
}

function test() {
  var target = TestRunner.mainTarget;

  function uiSourceCodeURL(uiSourceCode) {
    return uiSourceCode.url().replace(/.*LayoutTests/, 'LayoutTests');
  }

  function dumpUISourceCode(uiSourceCode, callback) {
    TestRunner.addResult('UISourceCode: ' + uiSourceCodeURL(uiSourceCode));
    if (uiSourceCode.contentType() === Common.resourceTypes.Script ||
        uiSourceCode.contentType() === Common.resourceTypes.Document)
      TestRunner.addResult(
          'UISourceCode is content script: ' +
          (uiSourceCode.project().type() === Workspace.projectTypes.ContentScripts));
    uiSourceCode.requestContent().then(didRequestContent);

    function didRequestContent(content, contentEncoded) {
      TestRunner.addResult('Highlighter type: ' + uiSourceCode.mimeType());
      TestRunner.addResult('UISourceCode content: ' + content);
      callback();
    }
  }

  TestRunner.runTestSuite([
    function testDocumentResource(next) {
      TestRunner.addResult('Creating resource.');
      TestRunner.waitForUISourceCode('resources/syntax-error.html').then(uiSourceCodeAdded);
      TestRunner.evaluateInPage('addIframe()');

      function uiSourceCodeAdded(uiSourceCode) {
        dumpUISourceCode(uiSourceCode, next);
      }
    },

    function testVMScript(next) {
      TestRunner.addResult('Creating script.');
      TestRunner.evaluateInPage('var foo=1;\n//# sourceURL=foo.js\n');
      TestRunner.waitForUISourceCode('foo.js').then(uiSourceCodeAdded);

      function uiSourceCodeAdded(uiSourceCode) {
        dumpUISourceCode(uiSourceCode, next);
      }
    },

    function testScriptResource(next) {
      TestRunner.addResult('Creating script resource.');
      TestRunner.evaluateInPage('addScript()');
      TestRunner.waitForUISourceCode('script1.js').then(uiSourceCodeAdded);

      function uiSourceCodeAdded(uiSourceCode) {
        dumpUISourceCode(uiSourceCode, next);
      }
    },

    function testRemoveStyleSheetFromModel(next) {
      TestRunner.waitForUISourceCode('style1.css').then(uiSourceCodeAdded);
      TestRunner.addResult('Creating stylesheet resource.');
      TestRunner.evaluateInPage(`addStyleSheet()`);

      function uiSourceCodeAdded(uiSourceCode) {
        TestRunner.addResult('Added uiSourceCode: ' + uiSourceCodeURL(uiSourceCode));
        TestRunner.waitForUISourceCodeRemoved(uiSourceCodeRemoved);
        TestRunner.evaluateInPage('removeStyleSheet()');
      }

      function uiSourceCodeRemoved(uiSourceCode) {
        TestRunner.addResult('Removed uiSourceCode: ' + uiSourceCodeURL(uiSourceCode));
        next();
      }
    }
  ]);
};
</script>
</head>
<body onload="runTest()">
<p>Tests NetworkUISourceCodeProvider class.</p>
</body>
</html>
