third_party/blink/web_tests/platform/generic/http/tests/devtools/sources/debugger/network-uisourcecode-provider-expected.txt

Tests NetworkUISourceCodeProvider class.


Running: testDocumentResource
Creating resource.
UISourceCode: http://127.0.0.1:8000/devtools/sources/debugger/resources/syntax-error.html
UISourceCode is content script: false
Highlighter type: text/html
UISourceCode content: <html>
<head>
<script>
)
</script>
</head>

<body onload="console.log('Iframe loaded');">
</body>
</html>

Running: testVMScript
Creating script.
UISourceCode: foo.js
UISourceCode is content script: false
Highlighter type: text/javascript
UISourceCode content: var foo=1;

Running: testScriptResource
Creating script resource.
UISourceCode: http://127.0.0.1:8000/devtools/sources/debugger/resources/script1.js
UISourceCode is content script: false
Highlighter type: text/javascript
UISourceCode content: function f1() {}

Running: testRemoveStyleSheetFromModel
Creating stylesheet resource.
Added uiSourceCode: http://127.0.0.1:8000/devtools/sources/debugger/resources/style1.css
Removed uiSourceCode: http://127.0.0.1:8000/devtools/sources/debugger/resources/style1.css
